refactor: extract useKeyboardShortcuts hook
