+++ src/App.jsx
@@ -68,6 +68,7 @@ import useSaveOnChange from './hooks/useSaveOnChange.js';
 import useTimelineScroll from './hooks/useTimelineScroll.js';
 import useModalClose from './hooks/useModalClose.js';
 import useMobileInteractions from './hooks/useMobileInteractions.js';
+import useKeyboardShortcuts from './hooks/useKeyboardShortcuts.js';
 
 // Encode a string that may contain non-ASCII characters as Base64.
 // btoa() throws InvalidCharacterError for codepoints > 255 (CJK, emoji, etc.).
@@ -1959,6 +1960,8 @@ const DayPlanner = () => {
     }
   }, [aiConfig]);
 
+  const enterFocusModeRef = useRef(null);
+
   const { longPressTriggeredRef, longPressTimerRef } = useMobileInteractions({
     isMobile, performUndo, performRedo,
   });
@@ -1987,208 +1990,29 @@ const DayPlanner = () => {
     showRecurrencePicker, setShowRecurrencePicker,
   });
 
-  // Global keyboard shortcuts
-  useEffect(() => {
-    const handleGlobalKeyDown = (e) => {
-      // Undo/Redo — works even when focus is in an input/textarea
-      if ((e.ctrlKey || e.metaKey) && !e.altKey) {
-        if (e.key === 'z' && !e.shiftKey) {
-          e.preventDefault();
-          performUndo();
-          return;
-        }
-        if ((e.key.toLowerCase() === 'z' && e.shiftKey) || e.key === 'y') {
-          e.preventDefault();
-          performRedo();
-          return;
-        }
-      }
-
-      // Cmd+K / Ctrl+K for spotlight search (works even in inputs)
-      if (e.key === 'k' && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        setShowSpotlight(prev => {
-          if (!prev) {
-            setSpotlightQuery('');
-            setSpotlightSelectedIndex(0);
-            playUISound('spotlight');
-          }
-          return !prev;
-        });
-        return;
-      }
-
-      // '?' for shortcut cheat sheet (works even in inputs)
-      if (e.key === '?' && !e.ctrlKey && !e.metaKey && !e.altKey) {
-        // Don't trigger in inputs unless it's already showing (to allow closing)
-        if (e.target.tagName === 'INPUT' || e.target.tagName === 'TEXTAREA' || e.target.isContentEditable) {
-          return;
-        }
-        e.preventDefault();
-        setShowShortcutHelp(prev => !prev);
-        return;
-      }
-
-      // Don't trigger if typing in an input or textarea
-      if (e.target.tagName === 'INPUT' || e.target.tagName === 'TEXTAREA' || e.target.isContentEditable) {
-        return;
-      }
-
-      // Don't trigger shortcuts when a modal is open (except Escape and ? handled above)
-      if (showAddTask || showFocusMode || showRoutinesDashboard || showShortcutHelp || showSpotlight || showSettings || showRemindersSettings || showWeeklyReview || showVoiceInput || showHabitModal || showFramesModal || frameAdjustModal || showRescheduleModal) {
-        return;
-      }
-
-      const noModifiers = !e.ctrlKey && !e.metaKey && !e.altKey;
-
-      // 'n' for new scheduled task
-      if (e.key === 'n' && noModifiers) {
-        e.preventDefault();
-        setNewTask({
-          title: '',
-          startTime: hoverPreviewTime || getNextQuarterHour(),
-          duration: 30,
-          date: hoverPreviewDate ? dateToString(hoverPreviewDate) : dateToString(selectedDate),
-          isAllDay: false,
-          openInInbox: false
-        });
-        setHoverPreviewTime(null);
-        setHoverPreviewDate(null);
-        setShowAddTask(true);
-      }
-
-      // 'i' for new inbox task
-      if (e.key === 'i' && noModifiers) {
-        e.preventDefault();
-        setNewTask({
-          title: '',
-          duration: 30,
-          openInInbox: true
-        });
-        setShowAddTask(true);
-      }
-
-      // 'r' for routines dashboard
-      if (e.key === 'r' && noModifiers && routinesEnabled) {
-        e.preventDefault();
-        setShowRoutinesDashboard(true);
-      }
-
-      // 'f' for focus mode (only when available)
-      if (e.key === 'f' && noModifiers) {
-        e.preventDefault();
-        if (focusModeAvailableRef.current) {
-          enterFocusMode();
-        }
-      }
-
-      // 'd' to toggle dark mode
-      if (e.key === 'd' && noModifiers) {
-        e.preventDefault();
-        setDarkMode(prev => !prev);
-      }
-
-      // 't' to jump to today
-      if (e.key === 't' && noModifiers) {
-        e.preventDefault();
-        goToToday();
-        if (showMonthView) {
-          const today = new Date();
-          setViewedMonth(new Date(today.getFullYear(), today.getMonth(), 1));
-        }
-      }
-
-      // 'm' to toggle month view
-      if (e.key === 'm' && noModifiers) {
-        e.preventDefault();
-        setShowMonthView(prev => !prev);
-      }
-
-      // '/' to toggle tag filter
-      if (e.key === '/' && noModifiers) {
-        e.preventDefault();
-        setShowMobileTagFilter(prev => !prev);
-      }
-
-      // 'b' to toggle backup menu
-      if (e.key === 'b' && noModifiers) {
-        e.preventDefault();
-        setShowBackupMenu(prev => !prev);
-      }
-
-      // ',' to switch side panel to Glance
-      if (e.key === ',' && noModifiers && !isMobile) {
-        e.preventDefault();
-        setTabletActiveTab('glance');
-      }
-
-      // '.' to switch side panel to Inbox
-      if (e.key === '.' && noModifiers && !isMobile) {
-        e.preventDefault();
-        setTabletActiveTab('inbox');
-      }
-
-      // 'v' for voice task input
-      if (e.key === 'v' && noModifiers && aiConfig.enabled && aiConfig.features.voiceTaskInput) {
-        e.preventDefault();
-        setShowVoiceInput(true);
-      }
-
-      // 'h' for habits modal
-      if (e.key === 'h' && noModifiers && habitsEnabled) {
-        e.preventDefault();
-        setShowHabitModal(true);
-      }
-
-      // 'e' for end-of-day reschedule
-      if (e.key === 'e' && noModifiers && aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0) {
-        e.preventDefault();
-        setShowRescheduleModal(true);
-        setRescheduleResults(null);
-        setRescheduleError('');
-      }
-
-      // 's' for smart schedule (open frames modal on schedule tab)
-      if (e.key === 's' && noModifiers && aiConfig?.enabled && aiConfig.features?.smartScheduling && gtdFrames.filter(f => f.enabled).length > 0) {
-        e.preventDefault();
-        if (isMobile) {
-          setMobileActiveTab('frames');
-          setFramesModalTab('schedule');
-          setEditingFrame(null);
-        } else {
-          setShowFramesModal(true);
-          setFramesModalTab('schedule');
-          setEditingFrame(null);
-        }
-      }
-
-      // Arrow left/right to navigate dates
-      if (e.key === 'ArrowLeft' && noModifiers) {
-        e.preventDefault();
-        changeDate(-1);
-        if (showMonthView) {
-          // Sync viewed month after date change
-          setSelectedDate(prev => {
-            setViewedMonth(new Date(prev.getFullYear(), prev.getMonth(), 1));
-            return prev;
-          });
-        }
-      }
-      if (e.key === 'ArrowRight' && noModifiers) {
-        e.preventDefault();
-        changeDate(1);
-        if (showMonthView) {
-          setSelectedDate(prev => {
-            setViewedMonth(new Date(prev.getFullYear(), prev.getMonth(), 1));
-            return prev;
-          });
-        }
-      }
-    };
-
-    document.addEventListener('keydown', handleGlobalKeyDown);
-    return () => document.removeEventListener('keydown', handleGlobalKeyDown);
-  }, [selectedDate, showAddTask, showShortcutHelp, showFocusMode, showRoutinesDashboard, showMonthView, showSpotlight, showSettings, showRemindersSettings, showWeeklyReview, showVoiceInput, showFramesModal, frameAdjustModal, showRescheduleModal, hoverPreviewTime, hoverPreviewDate, isMobile, routinesEnabled, habitsEnabled, aiConfig, gtdFrames]);
+  useKeyboardShortcuts({
+    performUndo, performRedo,
+    setShowSpotlight, setSpotlightQuery, setSpotlightSelectedIndex, playUISound,
+    setShowShortcutHelp,
+    showAddTask, showFocusMode, showRoutinesDashboard, showShortcutHelp, showSpotlight,
+    showSettings, showRemindersSettings, showWeeklyReview, showVoiceInput,
+    showHabitModal, showFramesModal, frameAdjustModal, showRescheduleModal,
+    selectedDate, hoverPreviewTime, hoverPreviewDate,
+    setNewTask, setShowAddTask, setHoverPreviewTime, setHoverPreviewDate,
+    routinesEnabled, setShowRoutinesDashboard,
+    focusModeAvailableRef, enterFocusModeRef,
+    setDarkMode,
+    showMonthView, goToToday, setViewedMonth,
+    setShowMonthView,
+    setShowMobileTagFilter,
+    setShowBackupMenu,
+    isMobile, setTabletActiveTab,
+    aiConfig, setShowVoiceInput,
+    habitsEnabled, setShowHabitModal,
+    gtdFrames, setShowRescheduleModal, setRescheduleResults, setRescheduleError,
+    setMobileActiveTab, setFramesModalTab, setEditingFrame, setShowFramesModal,
+    changeDate, setSelectedDate,
+  });
 
   const changeViewedMonth = (delta) => {
     setViewedMonth(prev => {
@@ -2689,6 +2513,7 @@ const DayPlanner = () => {
     // Android: immersive mode + pause notifications + DND
     nativeEnterFocusMode();
   };
+  enterFocusModeRef.current = enterFocusMode;
 
   const startFocusTimer = () => {
     setFocusShowSettings(false);
